+++ src/components/Accounts/Actions/FeeFineActions.js
@@ -6,7 +6,6 @@ import {
   FormattedMessage,
   injectIntl,
 } from 'react-intl';
-import SafeHTMLMessage from '@folio/react-intl-safe-html';
 import { stripesConnect } from '@folio/stripes/core';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import moment from 'moment';
@@ -245,7 +244,7 @@ class Actions extends React.Component {
     const fullName = getFullName(user);
 
     const message = (
-      <SafeHTMLMessage
+      <FormattedMessage
         id="ui-users.accounts.actions.cancellation.success"
         values={{
           count: 1,
@@ -576,7 +575,7 @@ class Actions extends React.Component {
       this.paymentStatus = paymentStatus;
 
       return (
-        <SafeHTMLMessage
+        <FormattedMessage
           id="ui-users.accounts.confirmation.message"
           values={{ count: 1, amount, action: paymentStatus }}
         />
@@ -593,7 +592,7 @@ class Actions extends React.Component {
       this.paymentStatus = paymentStatus;
 
       return (
-        <SafeHTMLMessage
+        <FormattedMessage
           id="ui-users.accounts.confirmation.message"
           values={{ count: accounts.length, amount, action: paymentStatus }}
         />